Removed the builtin module in interpreter as it was moved to proc/env.
