+++ my-scheme-interpreter/src/interpret.rs
@@ -1,4 +1,3 @@
-mod builtin;
 mod eval_helpers;
 pub mod interpreter;
 mod vm;
